refactor: move XXXTableName functions to dbutil pkg
