+++ internal/database/dbutil/util.go
@@ -38,3 +38,15 @@ func Fill(size int, elem string, sep string) string {
 
 	return strings.Join(r, sep)
 }
+
+func OfflineBatchTableName(groupID int, revisionID int64) string {
+	return fmt.Sprintf("offline_batch_%d_%d", groupID, revisionID)
+}
+
+func OfflineStreamSnapshotTableName(groupID int, revision int64) string {
+	return fmt.Sprintf("offline_stream_snapshot_%d_%d", groupID, revision)
+}
+
+func OfflineStreamCdcTableName(groupID int, revision int64) string {
+	return fmt.Sprintf("offline_stream_cdc_%d_%d", groupID, revision)
+}
